Tar only for APT
