+++ demo/src/main/edu/georgetown/library/fileAnalyzer/filetest/DemoActionRegistry.java
@@ -20,8 +20,7 @@ public class DemoActionRegistry extends DSpaceActionRegistry {
         add(new CreateAPTrustBag(dt)); 
 		add(new VerifyBag(dt)); 
 		add(new VerifyBagZip(dt)); 
-        add(new VerifyAPTrustBag(dt)); 
-        add(new VerifyAPTrustBagZip(dt)); 
+        add(new VerifyAPTrustBagTar(dt)); 
 		removeFT(CounterValidation.class);
 		add(new CounterValidationXls(dt)); 
         add(new DemoFileTest(dt));
